# server/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Ag Brain API
  version: 1.0.0
  description: Admin-only rural management API.
servers:
  - url: http://localhost:3333
paths:
  /health:
    get:
      security: []
      responses:
        "200": { description: Database is ready }
        "503": { description: Database is unavailable }
  /api/v1/auth/login:
    post:
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginInput" }
      responses:
        "200":
          description: Authenticated; returns the HttpOnly session cookie.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionResponse" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/auth/me:
    get:
      responses:
        "200":
          description: Current administrator
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/auth/session:
    delete:
      responses:
        "204": { description: Session deleted }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/producers:
    get:
      parameters:
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/Limit" }
        - { name: search, in: query, schema: { type: string } }
        - {
            name: status,
            in: query,
            schema: { $ref: "#/components/schemas/ProducerStatus" },
          }
      responses:
        "200": { description: Paginated producers }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateProducerInput" }
      responses:
        "201": { description: Producer created }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/BusinessRuleError" }
  /api/v1/producers/{id}:
    parameters:
      - {
          name: id,
          in: path,
          required: true,
          schema: { type: string, format: uuid },
        }
    get:
      responses:
        "200": { description: Producer }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProducerInput" }
      responses:
        "200": { description: Producer updated }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/BusinessRuleError" }
    delete:
      responses:
        "204": { description: Producer deleted }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/farms:
    get:
      parameters:
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/Limit" }
        - {
            name: idProducer,
            in: query,
            schema: { type: string, format: uuid },
          }
        - { name: search, in: query, schema: { type: string } }
        - {
            name: state,
            in: query,
            schema: { type: string, minLength: 2, maxLength: 2 },
          }
      responses:
        "200": { description: Paginated farms }
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateFarmInput" }
      responses:
        "201": { description: Farm created }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/BusinessRuleError" }
  /api/v1/farms/{id}:
    parameters:
      - {
          name: id,
          in: path,
          required: true,
          schema: { type: string, format: uuid },
        }
    get:
      responses:
        "200": { description: Farm }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateFarmInput" }
      responses:
        "200": { description: Farm updated }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/BusinessRuleError" }
    delete:
      responses:
        "204": { description: Farm deleted }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/harvests:
    get:
      parameters:
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/Limit" }
        - {
            name: idProducer,
            in: query,
            schema: { type: string, format: uuid },
          }
        - { name: idFarm, in: query, schema: { type: string, format: uuid } }
        - {
            name: year,
            in: query,
            schema: { type: integer, minimum: 2000, maximum: 2100 },
          }
      responses:
        "200": { description: Paginated harvests }
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateHarvestInput" }
      responses:
        "201": { description: Harvest created }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/harvests/{id}:
    parameters:
      - {
          name: id,
          in: path,
          required: true,
          schema: { type: string, format: uuid },
        }
    get:
      responses:
        "200": { description: Harvest }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateHarvestInput" }
      responses:
        "200": { description: Harvest updated }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      responses:
        "204": { description: Harvest deleted }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/audits:
    get:
      summary: List audit log entries
      parameters:
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/Limit" }
        - {
            name: operation,
            in: query,
            schema: { $ref: "#/components/schemas/AuditOperation" },
          }
        - {
            name: resource,
            in: query,
            schema: { $ref: "#/components/schemas/AuditResource" },
          }
        - {
            name: outcome,
            in: query,
            schema: { $ref: "#/components/schemas/AuditOutcome" },
          }
        - { name: idActor, in: query, schema: { type: string, format: uuid } }
        - {
            name: from,
            in: query,
            description: "Inclusive UTC day start (YYYY-MM-DD)",
            schema: { type: string, format: date },
          }
        - {
            name: to,
            in: query,
            description: "Exclusive next-day UTC start (YYYY-MM-DD)",
            schema: { type: string, format: date },
          }
        - { name: search, in: query, schema: { type: string, maxLength: 160 } }
      responses:
        "200":
          description: Paginated audit log entries newest first
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuditListResponse" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/dashboard:
    get:
      parameters:
        - {
            name: idProducer,
            in: query,
            schema: { type: string, format: uuid },
          }
        - {
            name: year,
            in: query,
            schema: { type: integer, minimum: 2000, maximum: 2100 },
          }
      responses:
        "200": { description: General or producer-filtered dashboard }
        "404": { $ref: "#/components/responses/NotFound" }
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: agbrain-session
  parameters:
    Page:
      {
        name: page,
        in: query,
        schema: { type: integer, minimum: 1, default: 1 },
      }
    Limit:
      {
        name: limit,
        in: query,
        schema: { type: integer, minimum: 1, maximum: 100, default: 20 },
      }
  responses:
    Unauthorized:
      description: Missing or invalid session
      content:
        { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    ValidationError:
      description: Invalid HTTP input
      content:
        { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: Resource not found
      content:
        { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Conflict:
      description: Unique resource conflict
      content:
        { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    BusinessRuleError:
      description: Business rule violation
      content:
        { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  schemas:
    Error:
      type: object
      required: [error, requestId]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: array, items: {} }
        requestId: { type: string, format: uuid }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8, maxLength: 128 }
    SessionResponse:
      type: object
      properties:
        user:
          type: object
          properties:
            idUser: { type: string, format: uuid }
            name: { type: string }
            email: { type: string, format: email }
    DocumentType: { type: string, enum: [CPF, CNPJ] }
    ProducerStatus: { type: string, enum: [ACTIVE, INACTIVE] }
    CreateProducerInput:
      type: object
      required: [name, documentType, document, email, city, state]
      properties:
        name: { type: string, maxLength: 160 }
        documentType: { $ref: "#/components/schemas/DocumentType" }
        document:
          {
            type: string,
            description: CPF or numeric/alphanumeric CNPJ,
            masked or canonical,
          }
        email: { type: string, format: email }
        city: { type: string, maxLength: 120 }
        state: { type: string, minLength: 2, maxLength: 2 }
        status: { $ref: "#/components/schemas/ProducerStatus" }
    UpdateProducerInput:
      type: object
      properties:
        name: { type: string, maxLength: 160 }
        documentType: { $ref: "#/components/schemas/DocumentType" }
        document:
          {
            type: string,
            description: CPF or numeric/alphanumeric CNPJ,
            masked or canonical,
          }
        email: { type: string, format: email }
        city: { type: string, maxLength: 120 }
        state: { type: string, minLength: 2, maxLength: 2 }
        status: { $ref: "#/components/schemas/ProducerStatus" }
    CreateFarmInput:
      type: object
      required:
        [
          idProducer,
          name,
          city,
          state,
          totalArea,
          agriculturalArea,
          vegetationArea,
        ]
      properties:
        idProducer: { type: string, format: uuid }
        name: { type: string, maxLength: 160 }
        city: { type: string, maxLength: 120 }
        state: { type: string, minLength: 2, maxLength: 2 }
        totalArea: { type: number, exclusiveMinimum: 0 }
        agriculturalArea: { type: number, minimum: 0 }
        vegetationArea: { type: number, minimum: 0 }
    UpdateFarmInput:
      type: object
      properties:
        name: { type: string, maxLength: 160 }
        city: { type: string, maxLength: 120 }
        state: { type: string, minLength: 2, maxLength: 2 }
        totalArea: { type: number, exclusiveMinimum: 0 }
        agriculturalArea: { type: number, minimum: 0 }
        vegetationArea: { type: number, minimum: 0 }
    CreateHarvestInput:
      type: object
      required: [idFarm, year, crops]
      properties:
        idFarm: { type: string, format: uuid }
        year: { type: integer, minimum: 2000, maximum: 2100 }
        crops:
          {
            type: array,
            minItems: 1,
            maxItems: 30,
            items: { type: string, maxLength: 100 },
          }
    UpdateHarvestInput:
      type: object
      properties:
        idFarm: { type: string, format: uuid }
        year: { type: integer, minimum: 2000, maximum: 2100 }
        crops:
          {
            type: array,
            minItems: 1,
            maxItems: 30,
            items: { type: string, maxLength: 100 },
          }
    AuditOperation:
      type: string
      enum:
        - auth.login
        - auth.logout
        - producer.create
        - producer.update
        - producer.delete
        - farm.create
        - farm.update
        - farm.delete
        - harvest.create
        - harvest.update
        - harvest.delete
    AuditResource:
      type: string
      enum: [SESSION, PRODUCER, FARM, HARVEST]
    AuditOutcome:
      type: string
      enum: [SUCCESS, FAILURE]
    AuditLog:
      type: object
      required:
        [
          idAudit,
          operation,
          resource,
          outcome,
          statusCode,
          metadata,
          occurredAt,
        ]
      properties:
        idAudit: { type: string, format: uuid }
        operation: { $ref: "#/components/schemas/AuditOperation" }
        resource: { $ref: "#/components/schemas/AuditResource" }
        idResource: { type: string, format: uuid, nullable: true }
        outcome: { $ref: "#/components/schemas/AuditOutcome" }
        idActor: { type: string, format: uuid, nullable: true }
        actorName: { type: string, nullable: true }
        actorEmail: { type: string, format: email, nullable: true }
        statusCode: { type: integer }
        errorCode: { type: string, nullable: true }
        errorMessage: { type: string, nullable: true }
        requestId: { type: string, nullable: true }
        ipAddress: { type: string, nullable: true }
        userAgent: { type: string, nullable: true }
        metadata: { type: object }
        occurredAt: { type: string, format: date-time }
    AuditListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/AuditLog" }
        meta:
          type: object
          required: [page, limit, total, lastPage]
          properties:
            page: { type: integer }
            limit: { type: integer }
            total: { type: integer }
            lastPage: { type: integer }

security:
  - cookieAuth: []